deleted grpc_client and moved newClient to the server
